refactor : controller에서  numberGenerator 필드 제거
- RandomNumberGenerator -> NumberGenerator로 이름 변경
- LottoNumberGenerator -> RandomNumberGenerator로 이름 변경
