+++ src/main/java/LottoApplication.java
@@ -1,6 +1,4 @@
 import controller.LottoController;
-import model.LottoNumberGenerator;
-import model.RandomNumberGenerator;
 import view.InputView;
 import view.OutputView;
 
@@ -9,8 +7,7 @@ public class LottoApplication {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
-        RandomNumberGenerator randomNumberGenerator = new LottoNumberGenerator();
 
-        new LottoController(inputView, outputView, randomNumberGenerator).run();
+        new LottoController(inputView, outputView).run();
     }
 }
