+++ src/main/java/model/NumberGenerator.java
@@ -1,0 +1,7 @@
+package model;
+
+import java.util.Set;
+
+public interface NumberGenerator {
+    Set<Integer> generateNumbers();
+}
