+++ src/test/java/model/RandomNumberGeneratorTest.java
@@ -6,13 +6,13 @@ import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class LottoNumberGeneratorTest {
+class RandomNumberGeneratorTest {
 
     @DisplayName("1~45사이의 6개의 무작위 정수를 생성한다")
     @Test
     void lottoGenerateTest() {
-        LottoNumberGenerator lottoNumberGenerator = new LottoNumberGenerator();
-        Set<Integer> numbers = lottoNumberGenerator.generateNumbers();
+        RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
+        Set<Integer> numbers = randomNumberGenerator.generateNumbers();
 
         assertThat(numbers.size()).isEqualTo(6);
         for (int number : numbers) {
